Use `SharedUrl` in more places
